+++ test/IO/test_io.py
@@ -93,6 +93,7 @@ binary_example_file_list = _convert_filelist([
     'N46E013.hgt',
     'example.zon',
     'example.nc',
+    'example.vk4',
     'mitutoyo_mock.xlsx',
     'mitutoyo_nonuniform_mock.xlsx'
 ] + [] if NuMPI._has_mpi4py else ['example-2d.npy'])  # MPI I/O does not support Python streams
@@ -577,6 +578,7 @@ def test_detect_format(file_format_examples):
     assert detect_format(os.path.join(file_format_examples, 'example-2d.npy')) == 'npy'
     assert detect_format(os.path.join(file_format_examples, 'surface.2048x2048.h5')) == 'h5'
     assert detect_format(os.path.join(file_format_examples, 'example.zon')) == 'zon'
+    assert detect_format(os.path.join(file_format_examples, 'example.vk4')) == 'vk4'
     assert detect_format(os.path.join(file_format_examples, 'mitutoyo_mock.xlsx')) == 'mitutoyo'
     assert detect_format(os.path.join(file_format_examples, 'mitutoyo_nonuniform_mock.xlsx')) == 'mitutoyo'