+++ SurfaceTopography/ChangeLog.md
@@ -7,8 +7,10 @@ v1.1 (not yet released)
 - API: `scale_dependent_statistical_property` of a `SurfaceContainer` now
   returns either `np.ndarray` or `np.ma.masked_array`
 - ENH: Reader for the Mitutoyo SurfTest Excel spread sheet format
+- ENH: Reader for Keyence VK4
 - MAINT: Added contact.engineering paper to bibliography
 - BUG: Check for reentrant topographies in slope calculation
+- BUG: Fixed aspect ratio used in convenience plotting function
 
 v1.0 (06May22)
 --------------